- deleting files (video, background, headshots)
 - add 4 textInputs for "Sport and Years played"

+++ csu_admin/src/com/blogspot/jaggerm/cdmeyer_admin/model/Athlete.as
@@ -1,5 +1,7 @@
 package com.blogspot.jaggerm.cdmeyer_admin.model
 {
+	import com.blogspot.jaggerm.cdmeyer_admin.utils.UDate;
+	
 	import flash.filesystem.File;
 	import flash.filesystem.FileStream;
 	import flash.net.FileReference;
@@ -7,7 +9,7 @@ package com.blogspot.jaggerm.cdmeyer_admin.model
 	
 	import mx.controls.Alert;
 	import mx.utils.StringUtil;
-
+	
 	public class Athlete
 	{
 		public var id : String = '';
@@ -22,6 +24,9 @@ package com.blogspot.jaggerm.cdmeyer_admin.model
 		public var bio : String = '';
 		public var headShots : Array = [];
 		public var oacs : Array = [];
+		
+		private var deleteVideos : Array = [];
+		private var deleteHeadShots : Array = [];
 		
 		public function Athlete(rawData : XML = null)
 		{
@@ -61,6 +66,13 @@ package com.blogspot.jaggerm.cdmeyer_admin.model
 			return info;
 		}
 		
+		public function reset():void
+		{
+			deleteVideos = [];
+			deleteHeadShots = [];
+		}
+		
+		
 		public function GetOacs() : void
 		{
 			oacs.splice(0, oacs.length);
@@ -73,8 +85,8 @@ package com.blogspot.jaggerm.cdmeyer_admin.model
 		
 		public function SaveOacs() : void
 		{
-//			if(oacs.length == 0)
-//				return;
+			//			if(oacs.length == 0)
+			//				return;
 			
 			var sportsXML : String = '<items>';
 			for each(var item : String in oacs)
@@ -101,11 +113,20 @@ package com.blogspot.jaggerm.cdmeyer_admin.model
 			var videosXML : String = '<items>';
 			for each(var item : VideoVO in videos)
 			{
-				videosXML += item.toXMLString();
+				if(item)
+					videosXML += item.toXMLString();
 			}
 			videosXML += '</items>';
 			
 			csu_admin.SaveFileContent(infoPath() + 'video.xml', videosXML);
+			
+			for each(item in deleteVideos)
+			{
+				var file:File = new File(csu_admin.APP_PATH + infoPath() + item.file);
+				if(file.exists)
+					file.deleteFile();
+			}
+			
 		}
 		
 		public function GetSports2() : void
@@ -152,72 +173,112 @@ package com.blogspot.jaggerm.cdmeyer_admin.model
 			var files : Array = file.getDirectoryListing();
 			for each(var item : File in files)
 			{
-				if((item.nativePath.toLowerCase().indexOf('.jpg') != -1) || (item.nativePath.toLowerCase().indexOf('.png') != -1))
+				if( ((item.nativePath.toLowerCase().indexOf('.jpg') != -1) || (item.nativePath.toLowerCase().indexOf('.png') != -1)) && item.name != backGroundImage)
 					headShots.push(item.name);
 			}
 		}
 		
 		public function SaveBackground(value : String) : void
 		{
+			deleteBackground();
+			
 			var tmp : Array = value.split('\\');
 			var file : File = new File(value);
 			backGroundImage = String(tmp[tmp.length-1]).replace(' ', '');
-//			Alert.show(backGroundImage);
-//			return;
+			var index:int = backGroundImage.lastIndexOf(".");
+			backGroundImage = backGroundImage.substring(0, index) + "_" + UDate.now() + backGroundImage.substr(index, backGroundImage.length);
 			var destination:FileReference = File.documentsDirectory.resolvePath(csu_admin.APP_PATH + infoPath() + backGroundImage);						
 			file.copyTo(destination,true);
+		}
+		
+		public function deleteBackground():void
+		{
+			var file:File = new File(csu_admin.APP_PATH + infoPath() + backGroundImage);
+			if(file.exists && backGroundImage)
+				file.deleteFile();
+			
+			backGroundImage = "";
 		}
 		
 		public function SaveHeadShot(id : String, path : String) : void
 		{
 			var tmp : Array = path.split('\\');
 			var file : File = new File(path);
-			var headShotImage : String = String(tmp[tmp.length-1]).replace(' ', '');
-			var destination:FileReference = File.documentsDirectory.resolvePath(csu_admin.APP_PATH + infoPath() + headShotImage);						
+			var fileName : String = String(tmp[tmp.length-1]).replace(' ', '');
+			var index:int = fileName.lastIndexOf(".");
+			fileName = fileName.substring(0, index) + "_" + UDate.now() + fileName.substr(index, fileName.length);
+			var destination:FileReference = File.documentsDirectory.resolvePath(csu_admin.APP_PATH + infoPath() + fileName);						
 			file.copyTo(destination,true);
 			
-			var index : uint = Number(id);
+			index = Number(id);
+			if(headShots.length > index)
+				headShots[index] = fileName;
+			else
+				headShots.push(fileName);
+		}
+		
+		public function deleteHeadShot(index:int):void
+		{
 			if(headShots.length > index)
 			{
-				file = new File(csu_admin.APP_PATH + infoPath() + headShots[index]);
-				if(file.exists)
-					file.deleteFile();
-				
-				headShots[index] = headShotImage;
+				deleteHeadShots.push(headShots[index]);
+				headShots[index] = null;
 			}
 			else
-				headShots.push(headShotImage);
+			{
+				deleteHeadShots.push( headShots.pop() );
+			}
+		}
+		
+		public function cleanHeadShots():void
+		{
+			for each(var fileName:String in deleteHeadShots)
+			{
+				var file:File = new File(csu_admin.APP_PATH + infoPath() + fileName);
+				if(file.exists)
+					file.deleteFile();
+			}
 		}
 		
 		public function SaveVideo(id : String, path : String, title : String) : void
 		{
 			var tmp : Array = path.split('\\');
 			var file : File = new File(path);
-			var headShotImage : String = String(tmp[tmp.length-1]).replace(' ', '');
-			var destination:FileReference = File.documentsDirectory.resolvePath(csu_admin.APP_PATH + infoPath() + headShotImage);						
+			var fileName : String = String(tmp[tmp.length-1]).replace(' ', '');
+			var index:int = fileName.lastIndexOf(".");
+			fileName = fileName.substring(0, index) + "_" + UDate.now() + fileName.substr(index, fileName.length);
+			var destination:FileReference = File.documentsDirectory.resolvePath(csu_admin.APP_PATH + infoPath() + fileName);						
 			file.copyTo(destination,true);
 			
 			var videoVO : VideoVO = new VideoVO();
-			videoVO.file = headShotImage;
+			videoVO.file = fileName;
 			videoVO.title = title;
 			
-			var index : uint = Number(id);
+			index = Number(id);
 			if(videos.length > index)
-			{
-				file = new File(csu_admin.APP_PATH + infoPath() + videos[index].file);
-				if(file.exists)
-					file.deleteFile();
-								
 				videos[index] = videoVO;
-			}
 			else
 				videos.push(videoVO);
 		}
 		
+		public function deleteVideo(index:int):void
+		{
+			if(videos.length > index)
+			{
+				deleteVideos.push(videos[index]);
+				videos[index] = null;
+			}
+			else
+			{
+				deleteVideos.push( videos.pop() );
+			}
+		}
+		
+		
 		public function UpdateVideoTitle(id : String, title : String) : void
 		{
 			var index : uint = Number(id);
-			if(videos.length > index)
+			if(videos.length > index && videos[index])
 				videos[index].title = title;
 		}
